requirements.txt
aiobotocore
boto3
fastapi
gunicorn
python-dotenv
uvicorn